aiForEng/requirements.txt
python-telegram-bot
python-dotenv
openai
langchain
tiktoken
pandas
numpy
scipy